Moved L1 Module to DQMOffline

+++ HLTriggerOffline/Tau/python/Validation/HLTTauReferences_cfi.py
@@ -2,7 +2,7 @@ import FWCore.ParameterSet.Config as cms
 
 TauMCProducer  = cms.EDProducer("HLTTauMCProducer",
                               GenParticles  = cms.untracked.InputTag("source"),
-                              ptMinTau      = cms.untracked.double(15),
+                              ptMinTau      = cms.untracked.double(10),
                               ptMinMuon     = cms.untracked.double(2),
                               ptMinElectron = cms.untracked.double(5),
                               BosonID       = cms.untracked.int32(23),
